Make real Makefiles and restructured output
OK. Actually made some useable make(1) files. For now, all input files
are assumed to be txt2man(1) files in md/*.3fortran.man run

   build/scripts/txt2man -h

For instructions on how to use the txt2man(1) markdown language; or see
documention on the txt2man(1) github site.

The main thing to do to help is to put files in md/ that are suitable
input for this program that describe Fortran 2018 intrinsics.

Your build platform probably has to be *nix or CygWin.

To rebuild

   cd build
   make

so if you edit md/cos.* and then rebuild you should have a new
manpage in man/man3; a new man index for the man/ directory;
an HTML document in docs/ that has been properly indexed into
docs/BOOKS_FORTRAN.html and docs/index.html; and a demo program
in src/.

Still lots to think about; but keeping it "KISS" during the first
iterations.